backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain-openai==0.1.14
langchain-community==0.2.7
pdfplumber==0.11.1
pydantic==2.7.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
faiss-cpu==1.8.0
